move out util clases for e2e tests.
