+++ MSAL/test/integration/native_auth/end_to_end/MSALNativeAuthEndToEndBaseTestCase.swift
@@ -25,71 +25,6 @@
 import XCTest
 import MSAL
 
-final class MSALNativeAuthEmailOTPUserPool {
-    enum ConfigurationError: LocalizedError, Equatable {
-        case missingOrEmptyValue(String)
-        case duplicateValues
-
-        var errorDescription: String? {
-            switch self {
-            case .missingOrEmptyValue(let key):
-                return "Email OTP username configuration '\(key)' is missing or empty."
-            case .duplicateValues:
-                return "Email OTP username configuration contains duplicate values."
-            }
-        }
-    }
-
-    private let usernames: [String]
-    private let lock = NSLock()
-    private var nextIndex = 0
-
-    init(usernames: [String]) {
-        precondition(!usernames.isEmpty)
-        self.usernames = usernames
-    }
-
-    static func make(configuration: [String: String], keys: [String]) throws -> MSALNativeAuthEmailOTPUserPool {
-        var usernames = [String]()
-
-        for key in keys {
-            guard let value = configuration[key]?.trimmingCharacters(in: .whitespacesAndNewlines),
-                  !value.isEmpty else {
-                throw ConfigurationError.missingOrEmptyValue(key)
-            }
-
-            usernames.append(value)
-        }
-
-        guard Set(usernames).count == usernames.count else {
-            throw ConfigurationError.duplicateValues
-        }
-
-        return MSALNativeAuthEmailOTPUserPool(usernames: usernames)
-    }
-
-    func nextUsername() -> String {
-        lock.lock()
-        defer { lock.unlock() }
-
-        let username = usernames[nextIndex]
-        nextIndex = (nextIndex + 1) % usernames.count
-        return username
-    }
-}
-
-enum MSALNativeAuthEmailOTPErrorClassifier {
-    private static let throttleErrorCode = 701014
-
-    static func isThrottleError(errorCodes: [Int], errorDescription: String?) -> Bool {
-        if errorCodes.contains(throttleErrorCode) {
-            return true
-        }
-
-        return errorDescription?.contains("AADSTS\(throttleErrorCode)") == true
-    }
-}
-
 class MSALNativeAuthEndToEndBaseTestCase: XCTestCase {
     private class Constants {
         static let nativeAuthKey = "native_auth"
@@ -126,7 +61,7 @@ class MSALNativeAuthEndToEndBaseTestCase: XCTestCase {
     // lifecycle. markCheckpoint() and the subsequent readOtpCode() run on the same instance within
     // a test, while parallel test runs never share mutable state.
     private let codeRetriever = MSALNativeAuthEmailCodeRetriever()
-    private var emailOTPUsername: String?
+    private let emailOTPUsernameProvider = MSALNativeAuthEmailOTPUsernameProvider()
     
     override class func setUp() {
         super.setUp()
@@ -261,15 +196,8 @@ class MSALNativeAuthEndToEndBaseTestCase: XCTestCase {
     }
 
     func emailOTPUsernameForCurrentTest() throws -> String {
-
-        if let emailOTPUsername {
-            return emailOTPUsername
-        }
-
         let userPool = try Self.configuredEmailOTPUserPool()
-        let username = userPool.nextUsername()
-        emailOTPUsername = username
-        return username
+        return emailOTPUsernameProvider.username(from: userPool)
     }
 
     func skipIfEmailOTPThrottled(
