+++ MSAL/test/integration/native_auth/end_to_end/sign_in/MSALNativeAuthSignInUsernameEndToEndTests.swift
@@ -27,79 +27,6 @@ import XCTest
 import MSAL
 
 final class MSALNativeAuthSignInUsernameEndToEndTests: MSALNativeAuthEndToEndBaseTestCase {
-    func test_emailOTPUserPoolRotatesAndWraps() {
-        let pool = MSALNativeAuthEmailOTPUserPool(usernames: ["user1", "user2", "user3"])
-
-        XCTAssertEqual(pool.nextUsername(), "user1")
-        XCTAssertEqual(pool.nextUsername(), "user2")
-        XCTAssertEqual(pool.nextUsername(), "user3")
-        XCTAssertEqual(pool.nextUsername(), "user1")
-    }
-
-    func test_emailOTPUserPoolRejectsInvalidConfiguration() {
-        XCTAssertThrowsError(
-            try MSALNativeAuthEmailOTPUserPool.make(
-                configuration: ["user1": "first", "user3": "third"],
-                keys: ["user1", "user2", "user3"]
-            )
-        ) { error in
-            XCTAssertEqual(
-                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
-                .missingOrEmptyValue("user2")
-            )
-        }
-
-        XCTAssertThrowsError(
-            try MSALNativeAuthEmailOTPUserPool.make(
-                configuration: ["user1": "first", "user2": "", "user3": "third"],
-                keys: ["user1", "user2", "user3"]
-            )
-        ) { error in
-            XCTAssertEqual(
-                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
-                .missingOrEmptyValue("user2")
-            )
-        }
-
-        XCTAssertThrowsError(
-            try MSALNativeAuthEmailOTPUserPool.make(
-                configuration: ["user1": "duplicate", "user2": "duplicate", "user3": "third"],
-                keys: ["user1", "user2", "user3"]
-            )
-        ) { error in
-            XCTAssertEqual(
-                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
-                .duplicateValues
-            )
-        }
-    }
-
-    func test_emailOTPUsernameIsRetainedForCurrentTest() throws {
-        let firstUsername = try emailOTPUsernameForCurrentTest()
-        let secondUsername = try emailOTPUsernameForCurrentTest()
-
-        XCTAssertEqual(firstUsername, secondUsername)
-    }
-
-    func test_emailOTPThrottleClassifierUsesErrorCodeAndDescription() {
-        XCTAssertTrue(MSALNativeAuthEmailOTPErrorClassifier.isThrottleError(errorCodes: [701014], errorDescription: nil))
-        XCTAssertTrue(
-            MSALNativeAuthEmailOTPErrorClassifier.isThrottleError(
-                errorCodes: [],
-                errorDescription: "AADSTS701014: Cannot generate more one time passcodes"
-            )
-        )
-    }
-
-    func test_emailOTPThrottleClassifierIgnoresUnrelatedErrors() {
-        XCTAssertFalse(
-            MSALNativeAuthEmailOTPErrorClassifier.isThrottleError(
-                errorCodes: [50034],
-                errorDescription: "AADSTS50034: User account does not exist"
-            )
-        )
-    }
-
     // Hero Scenario 2.2.1. Sign in - Use email and OTP to get token and sign in
     func test_signInAndSendingCorrectOTPResultsInSuccess() async throws {
         let username = try emailOTPUsernameForCurrentTest()
