+++ MSAL/test/unit/native_auth/utils/MSALNativeAuthEmailOTPUserPoolTests.swift
@@ -1,0 +1,85 @@
+//
+// Copyright (c) Microsoft Corporation.
+// All rights reserved.
+//
+// This code is licensed under the MIT License.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files(the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and / or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions :
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+
+import XCTest
+
+final class MSALNativeAuthEmailOTPUserPoolTests: XCTestCase {
+    func test_emailOTPUserPoolRotatesAndWraps() {
+        let pool = MSALNativeAuthEmailOTPUserPool(usernames: ["user1", "user2", "user3"])
+
+        XCTAssertEqual(pool.nextUsername(), "user1")
+        XCTAssertEqual(pool.nextUsername(), "user2")
+        XCTAssertEqual(pool.nextUsername(), "user3")
+        XCTAssertEqual(pool.nextUsername(), "user1")
+    }
+
+    func test_emailOTPUserPoolRejectsInvalidConfiguration() {
+        XCTAssertThrowsError(
+            try MSALNativeAuthEmailOTPUserPool.make(
+                configuration: ["user1": "first", "user3": "third"],
+                keys: ["user1", "user2", "user3"]
+            )
+        ) { error in
+            XCTAssertEqual(
+                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
+                .missingOrEmptyValue("user2")
+            )
+        }
+
+        XCTAssertThrowsError(
+            try MSALNativeAuthEmailOTPUserPool.make(
+                configuration: ["user1": "first", "user2": "", "user3": "third"],
+                keys: ["user1", "user2", "user3"]
+            )
+        ) { error in
+            XCTAssertEqual(
+                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
+                .missingOrEmptyValue("user2")
+            )
+        }
+
+        XCTAssertThrowsError(
+            try MSALNativeAuthEmailOTPUserPool.make(
+                configuration: ["user1": "duplicate", "user2": "duplicate", "user3": "third"],
+                keys: ["user1", "user2", "user3"]
+            )
+        ) { error in
+            XCTAssertEqual(
+                error as? MSALNativeAuthEmailOTPUserPool.ConfigurationError,
+                .duplicateValues
+            )
+        }
+    }
+
+    func test_emailOTPUsernameIsRetainedForCurrentTest() {
+        let pool = MSALNativeAuthEmailOTPUserPool(usernames: ["user1", "user2"])
+        let provider = MSALNativeAuthEmailOTPUsernameProvider()
+
+        let firstUsername = provider.username(from: pool)
+        let secondUsername = provider.username(from: pool)
+
+        XCTAssertEqual(firstUsername, secondUsername)
+        XCTAssertEqual(pool.nextUsername(), "user2")
+    }
+}
